quicktags ✅/source code/requirements.txt
PyQt5==5.15.9
Pillow==10.2.0
python-barcode==0.15.1
openpyxl==3.1.2
reportlab==4.1.0
pywin32==306       # Only if you use win32print or other Windows-specific logic